refactor(element): rename file "schedule" -> "scheduler"
